Added dynamic command specification loader/binder
Whew, what a commit. This implements a way to specify all commands and their metadata in a XML file (commands.xml), and load them at runtime. The loader will be able to construct all the necessary metadata to send to Discord for the command listing, and will also provide an interface to call the implementation of each command via reflection.